clm-pflotran: reenable checkpoint/restart of the clm-pflotran coupled model.
There are two outstanding issues:

1) Something has changed with waypoints, so the new waypoints that clm
is adding to pflotran are not always picked up correctly. When a clm
simulation starts after the final time specified in the pflotran input
file, pflotran will segfault in timestepper_base.F90 inside
TimestepperBaseSetTargetTime(). This is not strictly related to
checkpoint/restart, but will probably come up most frequently when
doing multiple clm restarts. The simple work around is to ensure the
final time in the pflotran input file is larger than the clm stop
time.

2) pflotran richards mode restarts should be bit for bit, but TH mode
is not currently bit for bit. This will take some investigation to
determine what's wrong. restart of TH modes should still be usable for
development, debugging and experimental simulations that don't require
bit for bit restarts.