bugfix: in clm-pflotran writing a restart file is causing a petsc error because the viewer is not initialized.
The problem appears to be that for surface-subsurface simulations with
subsurface only, pmc_base%is_master is not being set. This is set by
subsurface_factory.F90:SubsurfaceInitialize(), which in never called
by surface-subsurface, and is therefore unset by a subsurface only
simulation. This change sets is_master, but there may be other data
that is not initialized correctly. The bug appears to have been
introduced into clm-pflotran with changeset e79f1f6, but really needs
to be applied to pflotran-dev.