Added a safety check for the lower bound of the bisection search for the surface temperature from surface energy.
When we have small amounts of surface water, the amount of energy can go negative or correspond to an unrealistically small temperature (< -100 [C]). This causes the bisection method to fail as the root is no longer bracketed. This check simply tests for this condition, and if found leaves the surface temperature at its previous value. Modified regression gold files accordingly.