Fixes and improvements to the well_solver code follwoing testing:
Fix bug in the constuction of the well MPI communicator which caused the
global group rather than the well group to be used. This could hang on
some cross-proc well cases (although not the 2-proc ones in the
regression tests).

Fix bug in the wellbore oil density calculation. The getOilDensity
function does not return the oil phase molar density, as it name
suggests. Rather it returns the molar density of the oil component in
the oil, ignoring the dissolved gas. Add correction for this.

On the central wellbore composition solver, add a backtracker, which
limits the Newton iteration step if the error gets worse. This forces a
solve in one case in which the old version failed.

Add boolean voting system for ending convergence iterations for
cross-proc wells. This could prevent a hang if rounding error caused
slightly different residuals to be seen on different procs. Never
actually encountered, but possible.

Record and store MPI error returns.

Change to correct MPI types like MPI_Comm for MPI communicators and
groups.