This commit includes several changes that I've been trying as part of debugging efforts. It really should have been split up into several commits but the many rabbit holes I've been jumping into sort of messed up good coding practice. Here are the main changes in this commit:
1. There are many print to screen statements that need to be cleaned up
but they are in here unfortunately.
2. A final tweak in the equilibrium partitioning routine. I changed it
so that the aqueous concentration is pulled into the partitioning
routine rather than calculated before it goes into the routine. This
also required changing some of the constraint calculations, which made
some redundant code. But I figured redundant code was an okay compromise
since those calculations only happen once, in the beginning, in exchange
for more efficient code in the partitioning routine, which happens at
every Newton iteration.
3. Global mass balance has been implemented. It was useful for
debugging, so I finally did it and that's why it was done in the
debugging branch.
4. In the same spirit as (3), this commit also implements integral
fluxes, because it was also useful for debugging.
5. I played around with setting a minimum liquid saturation, but I'm not
sure if this is going to stay. However, it's in this commit, and it
might get cleaned up later.