A beta version of the geothermal fracture model is working. To summarize the functionality so far, this reads in a process model block within a TH mode simulation that looks like the following:
GEOTHERMAL_FRACTURE_MODEL

  MAX_DISTANCE 0.20 # [m]

  THERMAL_EXPANSION_COEFFICIENT 40.d-5 # [1/C] 40.d-6

  FRACTURE
    ID 1
    HYDRAULIC_APERTURE 1.d-2 # [m]
    CENTER 10.d0 0.5d0 4.5d0 # [m]
    NORMAL_VECTOR -0.25d0 0.d0 1.d0 # [m]
    RADIUS_X 100.d0 # [m]
    RADIUS_Y 100.d0 # [m]
  /
  FRACTURE
    ID 3
    HYDRAULIC_APERTURE 1.d-2 # [m]
    CENTER 10.d0 0.5d0 7.5d0 # [m]
    NORMAL_VECTOR -0.25d0 0.d0 1.d0 # [m]
    RADIUS_X 100.d0 # [m]
    RADIUS_Y 100.d0 # [m]
  /
  FRACTURE
    ID 2
    HYDRAULIC_APERTURE 6.d-3 # [m]
    CENTER 6.5d0 0.5d0 3.5d0 # [m]
    NORMAL_VECTOR 0.8d0 0.d0 1.d0 # [m]
    RADIUS_X 100.d0 # [m]
    RADIUS_Y 100.d0 # [m]
  /

/

Within an equivalent porous media domain, the cells intersected by a
fracture, and defined in the FRACTURE blocks, will be assigned an
anisotropic permeability following the methods in mapDFN. As the
temperature of the domain changes, thermal expansion/contraction
surrounding the fracture is calculated and the hydraulic aperature of
the fractures is modified. The permeability is re-calculated according
to the new hydraulic aperature, and the material auxvar is updated.

Development yet to be done:
- Make this work in parallel. This only works for a single process.
- Make the thermal expansion/contraction be an integral of length change
for some distance perpendicular to the fracture. It currently will only
consider the length change within the grid cell that the fracture
occupies.
- The fractures are assumed to be an infinite plane, but I need to
modify the geometric set up so that it can take in elipses, and
therefore use the RADIUS_X and RADIUS_Y keywords. Currently, these
keywords are ignored.