Beginning to add fracture family objects for automated dfn generation.
Fracture family fracture lists are now added to the global fracture
list, and the felxibility to use either individual fractures or fracture
family definitions is now functioning. The generation of fractures
within a fracture family is kludgy but the minimum things needed for the
set-up to run properly are there. Next, the actual statistical
distributions and sampling need to be implemented for the generation of
fractures in families.

Adds the infrastructure to sample the fracture family center from a
normal distribution around a mean value and standard deviation, along
with a seed for the random sampling method.

Added normal vector component sampling from a normal distribution but
there seems to be something wrong with the z coordinate sampling. If the
st dev for the z coordinate of the normal vector is anything but exact
zero, the permeability of the fracture goes NaN.

Fixed the issue for normal vector distribution sampling. I was missing
the step where I normalize the normal vector so that the arccos() isn't
undefined if the z component is outside of the range 1,-1.

Removed family probability and also trailing white space.

Added distribution and sampling for radius and aperture for fracture
families.