This branch was made directly from the top of jenn/wipp-well-nwt-srcsink-again, on 9JAN2024. In this commit, an optional keyword was added which allows the user to choose specific well segment numbers to be printed in the .well file. This has the benefit of reducing the size of the .well file output, if you are really only interested in seeing information on a few specific well segments rather than all of them. The way to invoke this capability in the input deck is like this:
PRINT_WELL_FILE

is changed to

PRINT_WELL_FILE SEGMENTS 2 6 14 18 31

Note that you must follow the PRINT_WELL_FILE with the new keyword
SEGMENTS, and then follow that with the list of segment numbers. An
error message will be spit out if no segment numbers are given, if a
segment number is invalid (meaning it is less than or equal to zero or
if it exceeds the number of well segments), and if the word SEGMENTS is
mispelled or some other word is given instead.